Updated the requirements for each os.
I could not manage to install with the current requirements files.
The linux one appeared to be a copy of the osx file. It contained builds
specific to osx and appnope which is osx specific. I have removed all
build details from this file and it now works.

Similarly for osx there were builds specifying pip installs which did
not work with conda. I changed these to be conda builds for the same
versions where possible.

I have added the instructions I had to use to create the environments to
the top of the files.